backend/requirements.txt
anyascii==0.3.2
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
babel==2.16.0
backports.functools-lru-cache==2.0.0
bcrypt==4.2.0
beautifulsoup4==4.13.4
boto3==1.38.1
botocore==1.38.1
Brotli==1.1.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.0
channels==4.1.0
charset-normalizer==3.3.2
click==8.1.8
constantly==23.10.4
cryptography==43.0.0
cssselect2==0.7.0
daphne==4.1.2
debugpy==1.3.0
defusedxml==0.7.1
Django==4.2.15
django-cors-headers==4.4.0
django-debug-toolbar==4.4.6
django-debugger==0.6
django-filter==24.3
django-google-sso==3.3.0
django-gst-field==0.0.3
django-model-utils==5.0.0
django-modelcluster==6.4
django-money==3.5.3
django-permissionedforms==0.1
django-render-block==0.10
django-request-logging==0.7.5
django-storages==1.14.6
django-stubs-ext==5.1.3
django-taggit==6.0.0
django-tasks==0.6.1
django-templated-email==3.0.1
django-treebeard==4.7.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
docopt==0.6.2
draftjs_exporter==5.1.0
et_xmlfile==2.0.0
filetype==1.2.0
fonttools==4.56.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
httplib2==0.22.0
hyperlink==21.0.0
idna==3.7
importlib_metadata==8.4.0
incremental==24.7.2
indic-transliteration==2.3.69
jmespath==1.0.1
laces==0.1.2
loguru==0.7.2
Markdown==3.7
markdown-it-py==3.0.0
mdurl==0.1.2
num2words==0.5.14
oauthlib==3.2.2
openpyxl==3.1.5
paramiko==3.5.0
pillow==11.0.0
pillow_heif==0.21.0
py-moneyed==3.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydyf==0.11.0
Pygments==2.19.1
PyJWT==2.9.0
PyNaCl==1.5.0
pyOpenSSL==24.2.1
pyparsing==3.1.4
pyphen==0.17.2
python-dateutil==2.9.0.post0
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rich==14.0.0
roman==5.0
rsa==4.9
s3transfer==0.12.0
service-identity==24.1.0
shellingham==1.5.4
six==1.17.0
soupsieve==2.7
sqlparse==0.5.1
telepath==0.3.1
tinycss2==1.4.0
tinyhtml5==2.0.0
toml==0.10.2
tomli==2.0.1
Twisted==24.7.0
txaio==23.1.1
typer==0.16.0
typing_extensions==4.12.2
urllib3==1.26.20
wagtail==6.4.1
weasyprint==64.1
webencodings==0.5.1
Willow==1.10.0
zipp==3.20.0
zope.interface==7.0.1
zopfli==0.2.3.post1
